% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_search_station.R
\name{dwd.search}
\alias{dwd.search}
\title{search DWD stations near you!}
\usage{
dwd.search(lon, lat, rad, ref)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{rad}{radius to search for nearby dwd stations}

\item{ref}{the reference period you want to compare the data to. This is important as it
then automatically filters out stations that don't have data in the time of the refernce period.
And no data means no data at all. It doesn't check for NAs. This is done in a later step
(see function montly.plot or dwd.cs.data)}
}
\value{
this function produces a table with the nearby (input: radius)
 found dwd station near by an entered (input) coordinate.
}
\description{
search dwd station by coordinate
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
