% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dwd_for_anomalies.R
\name{monthly.plot}
\alias{monthly.plot}
\title{FUNCTION_TITLE}
\usage{
monthly.plot(id, cnp, year, updateProgress)
}
\arguments{
\item{id}{of dwd station}

\item{cnp}{climate normal period (1 or 2: either 1961- 1990 or 1981 - 2010)}

\item{year}{the year of which I want to have the climate anomalies}

\item{updateProgress}{PARAM_DESCRIPTION}
}
\value{
a list of: aggregated and corrected data, the year (input) and the cnp (input)
}
\description{
getting dwd data of 1 year x to use in other functions
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
