% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cum_sum.R
\name{plot.cs}
\alias{plot.cs}
\title{FUNCTION_TITLE}
\usage{
\method{plot}{cs}(data_list)
}
\arguments{
\item{data_list}{list of dwd.cs.data function in get_dwd_for_cumsum.R script}
}
\value{
1 ggplot with x lines plus one reference line
}
\description{
cumsum plot of x years compared to reference period
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
}
