% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalies_precip.R
\name{precip.plot}
\alias{precip.plot}
\title{precipitation anomalies for a certain year}
\usage{
precip.plot(data_list)
}
\arguments{
\item{data_list}{list of monthly.plot function in get_dwd_for_anomalies.R script}
}
\value{
ggplot barplot of the year x of precipitation anomalies for year x and station x
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
