% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_cum_sum.R
\name{table.cs}
\alias{table.cs}
\title{FUNCTION_TITLE}
\usage{
table.cs(data_list)
}
\arguments{
\item{data_list}{list of dwd.cs.data function in get_dwd_for_cumsum.R script}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
percentage of rain that should have fallen of x years compared to the normal
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
}
