% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalies_temp.R
\name{temp.plot}
\alias{temp.plot}
\title{FUNCTION_TITLE}
\usage{
temp.plot(data_list)
}
\arguments{
\item{data_list}{list of monthly.plot function in get_dwd_for_anomalies.R script}
}
\value{
ggplot barplot of the year x of temperature anomalies for year x and station x
}
\description{
temperature anomalies for a certain year
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
